backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiosqlite
pydantic-settings
aiogram>=3.10,<4
google-genai
